+++ buttplug/src/core/message/client_device_message_attributes.rs
@@ -213,7 +213,7 @@ impl ClientDeviceMessageAttributesBuilder {
 pub struct NullDeviceMessageAttributes {}
 
 fn unspecified_feature() -> String {
-  "No description available for feature".to_string()
+  "N/A".to_string()
 }
 
 #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize, Getters, Setters)]